backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19

# CLIP and ML
torch>=2.0.0
torchvision>=0.15.0
transformers==4.47.1
Pillow==11.1.0
open-clip-torch==2.29.0

# Vector search
faiss-cpu==1.9.0.post1
numpy>=1.24.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1

# ML utilities
scikit-learn>=1.3.0
